Streamlit_Pages_Snowflake/requirements.txt
snowflake-connector-python
pandas
streamlit
matplotlib
python-dotenv